requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.1.0
django-redis==5.3.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
Faker==16.6.1
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
joblib==1.2.0
multidict==6.0.4
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.1
openpyxl==3.1.0
pandas==1.5.3
Pillow==9.4.0
PyJWT==2.7.0
pymemcache==4.0.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2022.7.1
qrcode==7.4.2
redis==4.5.5
regex==2022.10.31
requests==2.28.2
segno==1.5.2
six==1.16.0
sqlparse==0.4.3
tqdm==4.65.0
twilio==8.3.0
typing_extensions==4.6.3
tzdata==2022.7
urllib3==1.26.14
yarl==1.9.2
